pyproject.toml
[project]
name = "video-viral-agent"
version = "0.1.0"
description = "Multi-agent system to identify viral moments in videos using VLM and LLM"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "langchain-openrouter>=0.2.1",
    "moviepy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests", "src/__tests__"]
pythonpath = ["."]
